Online notification System
This is an e-notification system that helps you see and access notifications online. The basic idea is, for remote areas such as villages which won't have a good medium to transmit all the inforrmation and people still have to cisit the government office and inquire about basic details. Example scenario is,  peopledon't know the stock details in the ration distribution shop and  if governmenent is constructing a road,the details about when the road will be constructed and when it will end. It is still in development, any changes are welcomed.